+++ projects/parallel_tests/test_PAR_suspension.cpp
@@ -757,8 +757,8 @@ int main(int argc, char* argv[]) {
 #endif
 
     // Run simulation for specified time
-    int num_steps = std::ceil(time_end / time_step);
-    int out_steps = std::ceil((1 / time_step) / out_fps);
+    int num_steps = (int)std::ceil(time_end / time_step);
+    int out_steps = (int)std::ceil((1 / time_step) / out_fps);
     double time = 0;
     int out_frame = 0;
